move InjectAnnotationToContructRector test to correct place, rename FileReconstructor to FileProcessor
