--- tests/NodeVisitor/DependencyInjection/InjectAnnotationToConstructorRector/Test.php
@@ -1,16 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Rector\Tests\NodeVisitor\DependencyInjection\InjectAnnotationToConstructorRector;
-
-use Rector\Testing\PHPUnit\AbstractReconstructorTestCase;
-
-final class Test extends AbstractReconstructorTestCase
-{
-    public function test(): void
-    {
-        $this->doTestFileMatchesExpectedContent(
-            __DIR__ . '/wrong/wrong.php.inc',
-            __DIR__ . '/correct/correct.php.inc'
-        );
-    }
-}
