+++ tests/Rector/Contrib/Nette/InjectPropertyRector/correct/correct.php.inc
