+++ tests/Rector/Contrib/Nette/InjectPropertyRector/wrong/wrong.php.inc
